backend/requirements/base.txt
uvicorn
fastapi
python-multipart
pydantic
pydantic-settings
python-dotenv
sqlalchemy
python-jose
passlib
jinja2
openai